requirements.txt
numpy>=1.13.1
Pillow>=4.2.1
opencv>=3.1.0
pytorch>=0.1.12
torchvision>=0.1.8
scikit-learn>=0.19.0